networking: Fix review comments - move tests to separate files, refactoring
